add fetcher actor
