+++ Exercism/ExerciseList/ExerciseGridView.swift
@@ -77,7 +77,11 @@ struct ExerciseGridView: View {
 
 struct ExerciseGridView_Previews: PreviewProvider {
     static var previews: some View {
-        let modelData = TrackModel(client: ExercismClient(apiToken: Keys.token.rawValue), coordinator: AppCoordinator())
+        let token = ExercismKeychain.shared.get(for: Keys.token.rawValue)
+        let client = ExercismClient(apiToken: token!)
+        let fetcher = Fetcher(client: client)
+        let modelData = TrackModel(fetcher: fetcher,
+                                   coordinator: AppCoordinator())
         ExerciseGridView(exercise: modelData.exercises[0])
             .environmentObject(modelData)
     }
