+++ Exercism/Tracks/TrackGridView.swift
@@ -105,7 +105,11 @@ struct TrackGridView: View {
 
 struct TrackGridView_Previews: PreviewProvider {
     static var previews: some View {
-        let modelData = TrackModel(client: ExercismClient(apiToken: Keys.token.rawValue), coordinator: AppCoordinator())
+        let token = ExercismKeychain.shared.get(for: Keys.token.rawValue)
+        let client = ExercismClient(apiToken: token!)
+        let fetcher = Fetcher(client: client)
+        let modelData = TrackModel(fetcher: fetcher,
+                                   coordinator: AppCoordinator())
         TrackGridView(track: modelData.tracks[0])
             .environmentObject(modelData)
     }
